+++ system/system.h
@@ -1,8 +1,8 @@
 #ifndef SYSTEM_H
 #define SYSTEM_H
 
-float erlang_b_system(int channels, int lambda, float avg_duration, int n_samples);
-ErlangCstat erlang_c_system(int channels, int lambda, float avg_duration, int n_samples, float delay_threshold);
-ErlangGenStat erlang_gen_system(int channels, int lambda, float avg_duration, int n_samples, float delay_threshold, int queue_capacity);
+double erlang_b_system(int channels, int lambda, double avg_duration, int n_samples);
+ErlangCstat erlang_c_system(int channels, int lambda, double avg_duration, int n_samples, double delay_threshold);
+ErlangGenStat erlang_gen_system(int channels, int lambda, double avg_duration, int n_samples, double delay_threshold, int queue_capacity);
 
 #endif // SYSTEM_H
